requirements.txt
sqlalchemy==2.0.36
websocket-client==1.5.1
pandas==2.1.3
python-dotenv==1.0.1
python-binance==1.0.22
